refacor: move out all logic from app-component
