+++ react-components/src/components/ProductList/ProductList.css
@@ -5,5 +5,4 @@
   justify-content: space-between;
   flex-wrap: wrap;
   gap: 15px;
-  /* border: solid 1px blue; */
 }
