+++ react-components/src/components/context/appContext.tsx
@@ -19,8 +19,8 @@ export const AppProvider: React.FC<{ children: ReactNode }> =
   function AppProvider({ children }) {
     const [results, setResults] = useState<Product[]>([]);
     const [itemsTotalCount, setItemsTotalCount] = useState(0);
-    const [currentPage, setCurrentPage] = useState(1);
     const [itemsPerPage, setItemsPerPage] = useState(5);
+    const [currentPage, setCurrentPage] = useState(1);
 
     return (
       <AppContext.Provider
